added the jose package for jwt. updated component rendering for auth components created util dir in lib dir to hold new sessions.ts file and prisma.ts file was moved there as well
