+++ package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@prisma/client": "^6.7.0",
     "@prisma/extension-accelerate": "^1.3.0",
+    "jose": "^6.0.11",
     "jotai": "^2.12.4",
     "next": "15.3.1",
     "react": "^19.0.0",
