+++ src/lib/components/AuthComponent/SignUpComponent.tsx
@@ -57,6 +57,8 @@ function SignUpComponent() {
     <section id='section-signup' className={`transition-all ease-in-out duration-500 h-fit`} >
       {
         user != null ?
+          <div>Loading...</div>
+          :
           <>
             <button onClick={handleShowBtnClick} id='btn-show-login' className='transition-all ease-in-out duration-500 bg-blue-400 hover:bg-blue-600 active:bg-blue-200 px-2 py-1 rounded-full my-2 mx-auto w-fit text-sm tracking-widest font-semibold place-self-center cursor-pointer'>{
               showForm ?
@@ -70,8 +72,6 @@ function SignUpComponent() {
                 ""
             }
           </>
-          :
-          <div>Loading...</div>
       }
       
       {
